+++ test/cube_test.rb
@@ -100,8 +100,7 @@ class CubeTest < Minitest::Test
   def test_normalize
     item = CosineItem.new
     item.cube_embedding = [0, 3, 4]
-    # TODO use normalizes for Active Record 7.1+
-    assert_elements_in_delta [0, 3, 4], item.cube_embedding
+    assert_elements_in_delta [0, 0.6, 0.8], item.cube_embedding
     item.save!
     assert_elements_in_delta [0, 0.6, 0.8], item.cube_embedding
     assert_elements_in_delta [0, 0.6, 0.8], Item.last.cube_embedding
@@ -109,15 +108,13 @@ class CubeTest < Minitest::Test
 
   def test_insert
     CosineItem.insert!({cube_embedding: [0, 3, 4]})
-    # TODO use normalizes for Active Record 7.1+
-    expected = [0, 3, 4]
+    expected = supports_normalizes? ? [0, 0.6, 0.8] : [0, 3, 4]
     assert_elements_in_delta expected, Item.last.cube_embedding
   end
 
   def test_insert_all
     CosineItem.insert_all!([{cube_embedding: [0, 3, 4]}])
-    # TODO use normalizes for Active Record 7.1+
-    expected = [0, 3, 4]
+    expected = supports_normalizes? ? [0, 0.6, 0.8] : [0, 3, 4]
     assert_elements_in_delta expected, Item.last.cube_embedding
   end
 end