+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.5.0 (unreleased)
 
+- Changed `normalize` option to use Active Record normalization
 - Dropped support for Active Record < 7
 
 ## 0.4.3 (2024-09-02)